Add parallelization for matvec operations and improve performance
- Introduced a matvecPool to parallelize CSR's MulInto and MulTransposeInto operations across multiple goroutines, enhancing performance for large datasets.
- Updated the RecoverGroup and GroupZScore functions to utilize the new matvecPool for efficient matrix-vector multiplications.
- Added tests to verify the functionality and performance of the new parallelized operations.
- Adjusted wall time tests to account for race detector overhead in race builds.
- Refactored existing tests to improve clarity and maintainability.

+++ pkg/recover/csr_test.go
@@ -20,7 +20,7 @@ func buildTestCSR() *CSR {
 		NCols:  4,
 		RowPtr: []int32{0, 2, 3, 6},
 		ColIdx: []int32{0, 2, 1, 0, 1, 3},
-		Vals:   []float64{1.0, -2.0, 0.5, 1.0, 1.0, 3.0},
+		Vals:   []float32{1.0, -2.0, 0.5, 1.0, 1.0, 3.0},
 	}
 }
